fixes winners card on completed games
